requirements.txt
opencv-python
flask
pillow
imutils